// aircraft-maintenance-ui/src/app/services/models/page-response-auth-performed-work-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AuthPerformedWorkDto } from '../models/auth-performed-work-dto';
export interface PageResponseAuthPerformedWorkDto {
  content?: Array<AuthPerformedWorkDto>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}